Refactor download page components and add DownloadContent
- Simplified Jest configuration by merging object properties.
- Refactored the Download page to utilize a new DownloadContent component for better organization.
- Updated NavItem component to streamline link rendering.
- Introduced new CloudImage and DefaultImage components with improved structure and localization support.
- Enhanced table components with localized headers and cell content.
- Added internationalization support for various UI elements in the download section.
