+++ jest.config.mjs
@@ -1,8 +1,6 @@
 import nextJest from "next/jest.js";
 
-const createJestConfig = nextJest({
-  dir: "./",
-});
+const createJestConfig = nextJest({ dir: "./" });
 
 // Add any custom config to be passed to Jest
 /** @type {import('jest').Config} */
