+++ messages/en.json
@@ -25,9 +25,7 @@
         "sponsors": "Sponsors",
         "partners": "Partners"
       },
-      "resources": {
-        "gpgKeyInfo": "GPG Key Info"
-      },
+      "resources": { "gpgKeyInfo": "GPG Key Info" },
       "community": {
         "coc": "Code of Conduct",
         "forums": "Forums",
@@ -95,9 +93,7 @@
       "title": "Latest News",
       "description": "We regularly share updates with the community about what's happening. Here's the latest."
     },
-    "logoCloud": {
-      "title": "Backed By"
-    }
+    "logoCloud": { "title": "Backed By" }
   },
   "news": {
     "title": "News",
@@ -196,14 +192,8 @@
         "baseOs": "BaseOS Packages",
         "archived": "Archived Releases",
         "x86_64": {
-          "r9": {
-            "versionName": "Rocky Linux 9",
-            "plannedEol": "May 31 2032"
-          },
-          "r8": {
-            "versionName": "Rocky Linux 8",
-            "plannedEol": "May 31 2029"
-          }
+          "r9": { "versionName": "Rocky Linux 9", "plannedEol": "May 31 2032" },
+          "r8": { "versionName": "Rocky Linux 8", "plannedEol": "May 31 2029" }
         }
       },
       "cloudImages": {
@@ -219,15 +209,17 @@
               "region": "Region",
               "architecture": "Architecture",
               "imageId": "Image ID",
-              "deploy": "Deploy"
+              "deploy": "Deploy",
+              "search": "Search...",
+              "noResults": "No results.",
+              "previous": "Previous",
+              "next": "Next"
             }
           },
           "googleCloud": "Google Cloud",
           "azure": "Microsoft Azure"
         },
-        "downloadOptions": {
-          "qcow2": "QCOW2 Image"
-        }
+        "downloadOptions": { "qcow2": "QCOW2 Image" }
       },
       "container": {
         "title": "Docker (OCI) Images",
@@ -260,7 +252,7 @@
     "exportCompliance": {
       "title": "Export Compliance/Customs Information",
       "text1": "By downloading Rocky Linux software, you acknowledge that you understand all of the following:",
-      "text2": "Rocky Linux software and technical information may be subject to the U.S. Export Administration Regulations (the “EAR”) and other U.S. and foreign laws and may not be exported, re-exported or transferred (a) to a prohibited destination country under the EAR or U.S. sanctions regulations (currently Cuba, Iran, North Korea, Sudan, Syria, and the Crimea Region of Ukraine, subject to change as posted by the United States government); (b) to any prohibited destination or to any end user who has been prohibited from participating in U.S. export transactions by any federal agency of the U.S. government; or (c) for use in connection with the design, development or production of nuclear, chemical or biological weapons, or rocket systems, space launch vehicles, or sounding rockets, or unmanned air vehicle systems.",
+      "text2": "Rocky Linux software and technical information may be subject to the U.S. Export Administration Regulations (the \"EAR\") and other U.S. and foreign laws and may not be exported, re-exported or transferred (a) to a prohibited destination country under the EAR or U.S. sanctions regulations (currently Cuba, Iran, North Korea, Sudan, Syria, and the Crimea Region of Ukraine, subject to change as posted by the United States government); (b) to any prohibited destination or to any end user who has been prohibited from participating in U.S. export transactions by any federal agency of the U.S. government; or (c) for use in connection with the design, development or production of nuclear, chemical or biological weapons, or rocket systems, space launch vehicles, or sounding rockets, or unmanned air vehicle systems.",
       "narrative": "Rocky Linux in source code and binary code form is publicly available and is not subject to the EAR in accordance with §740.17(b)(1).",
       "product_title": "Product",
       "product_text": "Rocky Linux. Includes Rocky Linux Project \"Special Interest Group\" repositories.",
